api/requirements.txt
anyio==3.3.4
asgiref==3.4.1
asn1crypto==1.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
fastapi==0.70.0
google-api-core==2.2.1
google-auth==2.3.3
google-cloud-language==2.3.1
google-cloud-speech==2.11.1
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
h11==0.12.0
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
pg8000==1.22.0
proto-plus==1.19.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
scramp==1.4.1
shell==1.0.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
uvicorn==0.15.0
